// src/platform/packages/shared/shared-ux/code_editor/impl/editor.styles.d.ts
import type { EuiThemeComputed } from '@elastic/eui';
export declare const styles: {
    container: import("@emotion/react").SerializedStyles;
    keyboardHint: (euiTheme: EuiThemeComputed) => import("@emotion/react").SerializedStyles;
    controls: {
        base: (euiTheme: EuiThemeComputed) => import("@emotion/react").SerializedStyles;
        fullscreen: (euiTheme: EuiThemeComputed) => import("@emotion/react").SerializedStyles;
    };
};